refactor: extract browser, commands, app_core, input, and tui modules from main.rs (Phase 3) (#93)
Extract the remaining large chunks from main.rs into dedicated modules:
- app_core.rs: all core App impl methods (constructor, playback, history, etc.) + apply_sample_bounds
- browser.rs: browser App methods + list_browser_entries helpers
- commands.rs: execute_command App method + instr_editor_increment
- input.rs: full key event dispatch + per-view handlers (handle_input -> bool)
- tui.rs: run_tui with frame construction, status bar, modals, and event loop shell

main.rs is now ~1397 lines: module declarations, re-exports for test compatibility, main(), and the test suite.
