+++ vitakt/src/main.rs
@@ -1,19 +1,32 @@
 use anyhow::Result;
-use crossterm::{
-    event::{self, Event, KeyCode, KeyModifiers},
-    execute,
-    terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
+use ratatui::{style::Color, Terminal};
+use rtrb::RingBuffer;
+use std::{
+    path::PathBuf,
+    sync::{
+        atomic::{AtomicBool, AtomicU8, Ordering},
+        Arc,
+    },
 };
+use vitakt_core::{
+    audio::Command,
+    model::{ChainSlot, InterpMode, TRACKS},
+};
+
 mod app;
 use app::*;
+mod app_core;
 mod audio_stream;
 use audio_stream::start_audio_stream;
+mod browser;
+pub(crate) use browser::{list_browser_entries, list_browser_entries_ext};
 mod cli;
 use cli::{parse_args, CliAction};
+mod commands;
+pub(crate) use commands::instr_editor_increment;
 mod config;
-use config::Config;
 mod history;
-use history::History;
+mod input;
 mod note_utils;
 use note_utils::*;
 mod render;
@@ -21,30 +34,8 @@ use render::*;
 mod theme;
 #[cfg(test)]
 use theme::Theme;
+mod tui;
 mod wav_io;
-use wav_io::*;
-use ratatui::{
-    backend::CrosstermBackend,
-    layout::{Constraint, Direction, Layout, Rect},
-    style::{Color, Modifier, Style},
-    widgets::{Block, Borders, Clear, Paragraph},
-    Terminal,
-};
-use rtrb::RingBuffer;
-use std::{
-    io,
-    path::PathBuf,
-    sync::{
-        atomic::{AtomicBool, AtomicU32, AtomicU8, Ordering},
-        Arc,
-    },
-    time::Duration,
-};
-use vitakt_core::{
-    audio::Command,
-    model::{Chain, ChainSlot, FxCommand, InterpMode, Song, STEPS_PER_PHRASE, TRACKS},
-    storage,
-};
 
 // ── Shared test utilities ─────────────────────────────────────────────────────
 
@@ -53,2227 +44,6 @@ use vitakt_core::{
 /// don't race with each other (e.g. theme tests vs. config tests).
 #[cfg(test)]
 pub(crate) static HOME_MUTEX: std::sync::Mutex<()> = std::sync::Mutex::new(());
-
-// ── App state ─────────────────────────────────────────────────────────────────
-
-impl App {
-    fn new(
-        producer: Option<rtrb::Producer<Command>>,
-        sample_root: u8,
-        seq_playing: Arc<AtomicBool>,
-        current_seq_step: Arc<AtomicU8>,
-        preview_playing: Arc<AtomicBool>,
-    ) -> Self {
-        let song = Song::default(); // always has 1 phrase, 1 chain, 1 arrangement row
-        Self {
-            song,
-            view: View::SongView,
-            mode: InputMode::Normal,
-            cursor_step: 0,
-            octave: 4,
-            active_instrument: 0,
-            yy_pending: false,
-            yanked_step: None,
-            cmd_buf: String::new(),
-            status: "SONG  |  hjkl: nav  |  0-9: assign chain  |  Del: clear  |  Enter: chain view  |  SPC: play  |  q: quit".to_string(),
-            producer,
-            sample_root,
-            seq_playing,
-            current_seq_step,
-            instr_cursor: 0,
-            instr_editing: false,
-            instr_edit_buf: String::new(),
-            browser_entries: Vec::new(),
-            browser_cursor: 0,
-            browser_dir: std::env::current_dir().unwrap_or_else(|_| PathBuf::from(".")),
-            browser_mode: BrowserMode::Sample,
-            browser_scroll: 0,
-            startup_cursor: 0,
-            active_phrase_idx: 0,
-            view_stack: Vec::new(),
-            song_cursor_row: 0,
-            song_cursor_track: 0,
-            chain_view_track: 0,
-            chain_view_row: 0,
-            chain_cursor: 0,
-            chain_insert_mode: false,
-            cursor_col: 0,
-            fx_edit_buf: String::new(),
-            mixer_cursor_track: 0,
-            mixer_cursor_field: 0,
-            keyboard_instrument: 0,
-            render_receiver: None,
-            render_progress: Arc::new(AtomicU32::new(0)),
-            history: History::new(),
-            is_dirty: false,
-            is_previewing: false,
-            preview_playing,
-            status_timer: None,
-            theme: theme::load(),
-            config: Config::load(),
-            browser_show_bookmarks: false,
-            browser_bookmark_cursor: 0,
-        }
-    }
-
-    fn phrase_mut(&mut self) -> &mut vitakt_core::model::Phrase {
-        let idx = self.active_phrase_idx.min(self.song.phrases.len().saturating_sub(1));
-        &mut self.song.phrases[idx]
-    }
-
-    fn phrase(&self) -> &vitakt_core::model::Phrase {
-        let idx = self.active_phrase_idx.min(self.song.phrases.len().saturating_sub(1));
-        &self.song.phrases[idx]
-    }
-
-    /// Send a command to the audio thread (fire-and-forget).
-    fn send_cmd(&mut self, cmd: Command) {
-        if let Some(prod) = &mut self.producer {
-            let _ = prod.push(cmd);
-        }
-    }
-
-    /// Push a fresh phrase snapshot to the sequencer.
-    fn sync_phrase_to_sequencer(&mut self) {
-        let phrase = Box::new(self.phrase().clone());
-        self.send_cmd(Command::UpdatePhrase(phrase.clone()));
-        self.send_cmd(Command::UpdatePhraseInSong {
-            idx: self.active_phrase_idx,
-            phrase,
-        });
-        self.send_cmd(Command::SetSampleRoot(self.sample_root));
-    }
-
-    /// Send the full song data snapshot to the audio thread for arrangement playback.
-    fn sync_song_to_sequencer(&mut self) {
-        let arrangement = self.song.arrangement.clone();
-        let chains = self.song.chains.clone();
-        let phrases = self.song.phrases.clone();
-        let instruments = self.song.instruments.clone();
-        self.send_cmd(Command::UpdateSongData { arrangement, chains, phrases, instruments });
-        self.send_cmd(Command::SetSampleRoot(self.sample_root));
-    }
-
-    /// Push all mixer state to the audio thread (called after load and incremental changes).
-    fn sync_mixer_to_audio(&mut self) {
-        for t in 0..TRACKS {
-            let (vol, pan, mute, solo) = {
-                let m = &self.song.mixer[t];
-                (m.volume, m.pan, m.mute, m.solo)
-            };
-            self.send_cmd(Command::SetTrackVolume { track: t as u8, volume: vol });
-            self.send_cmd(Command::SetTrackPan { track: t as u8, pan });
-            self.send_cmd(Command::SetTrackMute { track: t as u8, mute });
-            self.send_cmd(Command::SetTrackSolo { track: t as u8, active: solo });
-        }
-    }
-
-    /// Toggle play / stop.
-    fn toggle_play(&mut self) {
-        if self.seq_playing.load(Ordering::Relaxed) {
-            self.send_cmd(Command::Stop);
-            self.seq_playing.store(false, Ordering::Relaxed);
-        } else {
-            self.sync_song_to_sequencer();
-            self.send_cmd(Command::Play);
-            self.seq_playing.store(true, Ordering::Relaxed);
-        }
-    }
-
-    /// Restart sequencer from step 0 (F5).
-    fn restart_play(&mut self) {
-        self.sync_song_to_sequencer();
-        self.send_cmd(Command::Restart);
-        self.seq_playing.store(true, Ordering::Relaxed);
-    }
-
-    /// Ensure instrument slots 0..=idx exist (creates defaults up to MAX_INSTRUMENTS).
-    fn ensure_instrument(&mut self, idx: usize) {
-        while self.song.instruments.len() <= idx && self.song.instruments.len() < MAX_INSTRUMENTS {
-            self.song.instruments.push(vitakt_core::model::Instrument::default());
-        }
-    }
-
-    /// Open the instrument editor for the active instrument, creating it if needed.
-    fn open_instrument_editor(&mut self) {
-        if self.song.instruments.len() < MAX_INSTRUMENTS {
-            self.ensure_instrument(self.active_instrument);
-        }
-        self.push_view(View::InstrumentEditor);
-        self.instr_cursor = 0;
-        self.instr_editing = false;
-        self.instr_edit_buf.clear();
-    }
-
-    /// Reload sample from disk for the active instrument and send a LoadVoice command.
-    fn reload_instrument_sample(&mut self) {
-        let idx = self.active_instrument;
-        if let Some(instr) = self.song.instruments.get(idx) {
-            if let Some(sample) = &instr.sample {
-                let path = sample.path.clone();
-                let embedded_bytes = sample.bytes.clone();
-                let loop_start = instr.loop_start.unwrap_or(0);
-                let loop_end = instr.loop_end.unwrap_or(0);
-                let sample_start = instr.sample_start;
-                let sample_end = instr.sample_end;
-                let interp_mode = instr.interp_mode.clone();
-                let load_result = if let Some(bytes) = embedded_bytes {
-                    load_wav_from_bytes(&bytes)
-                } else {
-                    load_wav(&path)
-                };
-                match load_result {
-                    Ok((buf, channels)) => {
-                        let (buf, loop_start, loop_end) =
-                            apply_sample_bounds(buf, channels, sample_start, sample_end, loop_start, loop_end);
-                        self.send_cmd(Command::LoadVoice {
-                            slot: 0,
-                            samples: buf,
-                            channels,
-                            loop_start,
-                            loop_end,
-                            interp_mode,
-                        });
-                        self.sample_root = self.song.instruments[idx].root_note;
-                        self.send_cmd(Command::SetSampleRoot(self.sample_root));
-                        self.status = format!("Loaded: {path}");
-                    }
-                    Err(e) => self.status = format!("Error: {e}"),
-                }
-            }
-        }
-    }
-
-    /// Reload instrument 0's sample after a project load.
-    fn reload_instruments(&mut self) {
-        if self.song.instruments.is_empty() {
-            return;
-        }
-        // Reload slot 0 (active instrument) if it has a sample path.
-        let instr = &self.song.instruments[0];
-        if let Some(sample) = &instr.sample {
-            let path = sample.path.clone();
-            let embedded_bytes = sample.bytes.clone();
-            let loop_start = instr.loop_start.unwrap_or(0);
-            let loop_end = instr.loop_end.unwrap_or(0);
-            let sample_start = instr.sample_start;
-            let sample_end = instr.sample_end;
-            let interp_mode = instr.interp_mode.clone();
-            let load_result = if let Some(bytes) = embedded_bytes {
-                load_wav_from_bytes(&bytes)
-            } else {
-                load_wav(&path)
-            };
-            match load_result {
-                Ok((buf, channels)) => {
-                    let (buf, loop_start, loop_end) =
-                        apply_sample_bounds(buf, channels, sample_start, sample_end, loop_start, loop_end);
-                    self.send_cmd(Command::LoadVoice {
-                        slot: 0,
-                        samples: buf,
-                        channels,
-                        loop_start,
-                        loop_end,
-                        interp_mode,
-                    });
-                    self.sample_root = self.song.instruments[0].root_note;
-                    self.send_cmd(Command::SetSampleRoot(self.sample_root));
-                }
-                Err(e) => self.status = format!("Warning: could not reload sample: {e}"),
-            }
-        }
-    }
-
-    /// Open the sample browser starting at the instrument's current sample directory,
-    /// or the current working directory if no sample is set.
-    fn open_sample_browser(&mut self) {
-        let start_dir = self
-            .song
-            .instruments
-            .get(self.active_instrument)
-            .and_then(|instr| instr.sample.as_ref())
-            .and_then(|sample| {
-                let p = std::path::Path::new(&sample.path);
-                p.parent().map(|parent| parent.to_path_buf())
-            })
-            .filter(|p| p.is_dir())
-            .unwrap_or_else(|| {
-                std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."))
-            });
-        self.browser_dir = start_dir;
-        self.browser_mode = BrowserMode::Sample;
-        self.browser_entries = list_browser_entries(&self.browser_dir);
-        self.browser_cursor = 0;
-        self.browser_scroll = 0;
-        self.push_view(View::SampleBrowser);
-    }
-
-    /// Open the file browser scoped to `.trk` project files.
-    fn open_project_browser(&mut self) {
-        let start_dir = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
-        self.browser_dir = start_dir;
-        self.browser_mode = BrowserMode::Project;
-        self.browser_entries = list_browser_entries_ext(&self.browser_dir, "trk");
-        self.browser_cursor = 0;
-        self.browser_scroll = 0;
-        self.push_view(View::SampleBrowser);
-    }
-
-    /// Handle Enter in the sample browser.
-    /// Navigates into a directory or loads a `.wav` file into the active instrument.
-    fn browser_enter(&mut self) {
-        if let Some(entry) = self.browser_entries.get(self.browser_cursor).cloned() {
-            match entry {
-                BrowserEntry::ParentDir => {
-                    self.browser_go_up();
-                }
-                BrowserEntry::Dir(name) => {
-                    self.browser_dir = self.browser_dir.join(&name);
-                    let ext = match self.browser_mode {
-                        BrowserMode::Sample => "wav",
-                        BrowserMode::Project => "trk",
-                    };
-                    self.browser_entries = list_browser_entries_ext(&self.browser_dir, ext);
-                    self.browser_cursor = 0;
-                    self.browser_scroll = 0;
-                }
-                BrowserEntry::Wav(name) => match self.browser_mode {
-                    BrowserMode::Sample => {
-                        let full_path = self.browser_dir.join(&name);
-                        let path_str = full_path
-                            .canonicalize()
-                            .unwrap_or(full_path)
-                            .to_string_lossy()
-                            .to_string();
-                        self.record("select sample");
-                        self.ensure_instrument(self.active_instrument);
-                        if let Some(instr) = self.song.instruments.get_mut(self.active_instrument) {
-                            instr.sample = Some(vitakt_core::model::Sample::from_path(path_str));
-                        }
-                        self.pop_view();
-                        self.reload_instrument_sample();
-                    }
-                    BrowserMode::Project => {
-                        let full_path = self.browser_dir.join(&name);
-                        let path_str = full_path
-                            .canonicalize()
-                            .unwrap_or(full_path)
-                            .to_string_lossy()
-                            .to_string();
-                        match storage::load_trk(&path_str) {
-                            Ok(song) => {
-                                self.song = vitakt_core::model::migrate(song);
-                                if self.song.phrases.is_empty() {
-                                    self.song
-                                        .phrases
-                                        .push(vitakt_core::model::Phrase::default());
-                                }
-                                self.sync_phrase_to_sequencer();
-                                self.reload_instruments();
-                                self.sync_song_to_sequencer();
-                                self.sync_mixer_to_audio();
-                                self.history.clear();
-                                // Clear nav stack and go straight to Song View.
-                                self.view_stack.clear();
-                                self.view = View::SongView;
-                                self.browser_mode = BrowserMode::Sample;
-                                self.set_timed_status(format!("Loaded: {path_str}"));
-                            }
-                            Err(e) => {
-                                self.set_timed_status(format!("Error: {e}"));
-                            }
-                        }
-                    }
-                },
-            }
-        }
-    }
-
-    /// Navigate to the parent directory in the sample browser.
-    fn browser_go_up(&mut self) {
-        if let Some(parent) = self.browser_dir.parent().map(|p| p.to_path_buf()) {
-            self.browser_dir = parent;
-            let ext = match self.browser_mode {
-                BrowserMode::Sample => "wav",
-                BrowserMode::Project => "trk",
-            };
-            self.browser_entries = list_browser_entries_ext(&self.browser_dir, ext);
-            self.browser_cursor = 0;
-            self.browser_scroll = 0;
-        }
-    }
-
-    /// Handle the `b` key in the sample browser: open the bookmark overlay if valid bookmarks
-    /// exist, otherwise show a hint in the status bar.
-    fn browser_try_open_bookmarks(&mut self) {
-        let has_valid = self
-            .config
-            .bookmarks
-            .iter()
-            .any(|p| std::path::Path::new(p.as_str()).is_dir());
-        if has_valid {
-            self.browser_bookmark_cursor = 0;
-            self.browser_show_bookmarks = true;
-        } else {
-            self.set_timed_status(
-                "No bookmarks set (or none exist on disk)  —  press B to add one".to_string(),
-            );
-        }
-    }
-
-    /// Handle the `B` key in the sample browser: append the current directory to
-    /// `config.bookmarks` (deduplicated) and persist.
-    fn browser_add_bookmark(&mut self) {
-        let dir = self.browser_dir.to_string_lossy().to_string();
-        if !self.config.bookmarks.contains(&dir) {
-            self.config.bookmarks.push(dir.clone());
-            if let Err(e) = self.config.save() {
-                self.set_timed_status(format!("Error saving bookmark: {e}"));
-            } else {
-                self.set_timed_status(format!("Bookmarked: {dir}"));
-            }
-        } else {
-            self.set_timed_status(format!("Already bookmarked: {dir}"));
-        }
-    }
-
-    /// Handle Enter on the bookmark overlay: navigate to the selected bookmark and close the
-    /// overlay.
-    fn browser_overlay_enter(&mut self) {
-        let valid_bookmarks: Vec<String> = self
-            .config
-            .bookmarks
-            .iter()
-            .filter(|p| std::path::Path::new(p.as_str()).is_dir())
-            .cloned()
-            .collect();
-        if let Some(path) = valid_bookmarks.get(self.browser_bookmark_cursor) {
-            let dest = PathBuf::from(path);
-            self.browser_dir = dest.clone();
-            self.browser_entries = list_browser_entries(&dest);
-            self.browser_cursor = 0;
-            self.browser_scroll = 0;
-        }
-        self.browser_show_bookmarks = false;
-        self.browser_bookmark_cursor = 0;
-    }
-
-    /// Handle Esc on the bookmark overlay: close without navigating.
-    fn browser_overlay_esc(&mut self) {
-        self.browser_show_bookmarks = false;
-    }
-
-    /// Adjust `browser_scroll` so that `browser_cursor` stays within the visible window.
-    /// `available` is the number of entry rows visible (after accounting for borders and header).
-    fn browser_clamp_scroll(&mut self, available: usize) {
-        if available == 0 {
-            return;
-        }
-        if self.browser_cursor < self.browser_scroll {
-            self.browser_scroll = self.browser_cursor;
-        } else if self.browser_cursor >= self.browser_scroll + available {
-            self.browser_scroll = self.browser_cursor - available + 1;
-        }
-    }
-
-    /// Handle Space in the sample browser: toggle preview playback of the highlighted .wav.
-    /// Silently ignores Space on a directory entry or an empty list.
-    fn browser_preview_toggle(&mut self) {
-        // Sync is_previewing from the shared atomic so stale state after natural
-        // completion doesn't cause a double-press to restart.
-        self.is_previewing = self.preview_playing.load(Ordering::Relaxed);
-
-        if self.is_previewing {
-            self.send_cmd(Command::StopPreview);
-            self.preview_playing.store(false, Ordering::Relaxed);
-            self.is_previewing = false;
-            return;
-        }
-        let Some(entry) = self.browser_entries.get(self.browser_cursor).cloned() else {
-            return;
-        };
-        let BrowserEntry::Wav(name) = entry else {
-            return; // directory — ignore
-        };
-        let full_path = self.browser_dir.join(&name);
-        let path_str = full_path
-            .canonicalize()
-            .unwrap_or(full_path)
-            .to_string_lossy()
-            .to_string();
-        match load_wav(&path_str) {
-            Ok((samples, channels)) => {
-                self.send_cmd(Command::PreviewSample { samples, channels });
-                self.preview_playing.store(true, Ordering::Relaxed);
-                self.is_previewing = true;
-            }
-            Err(e) => self.set_timed_status(format!("Preview error: {e}")),
-        }
-    }
-
-    /// Adjust BPM by `delta` and send the new value to the audio thread.
-    fn adjust_bpm(&mut self, delta: f32) {
-        self.record("set BPM");
-        self.song.bpm = (self.song.bpm + delta).clamp(20.0, 999.0);
-        self.send_cmd(Command::SetBpm(self.song.bpm));
-    }
-
-    /// Push the current view onto the navigation stack and switch to `next`.
-    fn push_view(&mut self, next: View) {
-        let current = std::mem::replace(&mut self.view, next);
-        self.view_stack.push(current);
-    }
-
-    /// Pop the navigation stack and return to the previous view.
-    fn pop_view(&mut self) {
-        if let Some(prev) = self.view_stack.pop() {
-            self.view = prev;
-        }
-        // Stop any active preview when leaving the sample browser.
-        if self.is_previewing {
-            self.send_cmd(Command::StopPreview);
-            self.is_previewing = false;
-            self.preview_playing.store(false, Ordering::Relaxed);
-        }
-        // Clear mode when returning to PhraseEditor
-        if matches!(self.view, View::PhraseEditor) {
-            self.mode = InputMode::Normal;
-        }
-        // Clear chain insert mode when leaving ChainView
-        self.chain_insert_mode = false;
-    }
-
-    /// Ensure chain slots exist up to and including `chain_idx`.
-    pub fn ensure_chain(&mut self, chain_idx: usize) {
-        while self.song.chains.len() <= chain_idx {
-            self.song.chains.push(Chain {
-                slots: vec![ChainSlot { phrase: 0, transpose: 0 }],
-            });
-        }
-    }
-
-    /// Enter a note at the current step and advance the cursor.
-    fn enter_note(&mut self, midi: u8) {
-        self.record(&format!("set note {} at step {}", note_name(midi), self.cursor_step));
-        let cursor = self.cursor_step;
-        let instr = self.active_instrument as u8;
-        let step = &mut self.phrase_mut().steps[cursor];
-        step.note = Some(midi);
-        step.instrument = Some(instr);
-        step.velocity = 100;
-
-        // Send NoteOn to audio thread for live preview.
-        if let Some(prod) = &mut self.producer {
-            let speed = pitch_speed(midi, self.sample_root);
-            let _ = prod.push(Command::NoteOn { slot: 0, speed });
-        }
-
-        // Keep sequencer phrase in sync.
-        self.sync_phrase_to_sequencer();
-
-        // Advance cursor.
-        self.cursor_step = (self.cursor_step + 1) % STEPS_PER_PHRASE;
-    }
-
-    fn execute_command(&mut self) {
-        let raw = self.cmd_buf.trim().to_string();
-        self.cmd_buf.clear();
-        self.mode = InputMode::Normal;
-
-        if let Some(path) = raw.strip_prefix("w ") {
-            let path = path.trim();
-            match storage::save_trk(&self.song, path) {
-                Ok(_) => {
-                    self.is_dirty = false;
-                    self.status = format!("Saved: {path}");
-                }
-                Err(e) => self.status = format!("Error: {e}"),
-            }
-        } else if let Some(path) = raw.strip_prefix("e ") {
-            let path = path.trim();
-            match storage::load_trk(path) {
-                Ok(song) => {
-                    self.song = vitakt_core::model::migrate(song);
-                    if self.song.phrases.is_empty() {
-                        self.song.phrases.push(vitakt_core::model::Phrase::default());
-                    }
-                    self.sync_phrase_to_sequencer();
-                    self.reload_instruments();
-                    self.sync_song_to_sequencer();
-                    self.sync_mixer_to_audio();
-                    self.history.clear();
-                    self.is_dirty = false;
-                    self.status = format!("Loaded: {path}");
-                }
-                Err(e) => self.status = format!("Error: {e}"),
-            }
-        } else if let Some(path) = raw.strip_prefix("export-json ") {
-            let path = path.trim();
-            match storage::export_json(&self.song, path) {
-                Ok(_) => self.status = format!("JSON exported: {path}"),
-                Err(e) => self.status = format!("Error: {e}"),
-            }
-        } else if let Some(path) = raw.strip_prefix("export-packed ") {
-            let path = path.trim().to_string();
-            self.status = format!("Packing samples…");
-            match storage::save_packed_trk(&self.song, &path) {
-                Ok(failed) if failed.is_empty() => {
-                    self.status = format!("Packed: {path}");
-                }
-                Ok(failed) => {
-                    self.status = format!(
-                        "Packed: {path} ({} sample(s) missing: {})",
-                        failed.len(),
-                        failed.join(", ")
-                    );
-                }
-                Err(e) => self.status = format!("Error: {e}"),
-            }
-        } else if let Some(path) = raw.strip_prefix("export-mix ") {
-            let path = path.trim().to_string();
-            self.start_render_mix(path);
-        } else if let Some(dir) = raw.strip_prefix("export-stems ") {
-            let dir = dir.trim().to_string();
-            self.start_render_stems(dir);
-        } else if let Some(rest) = raw.strip_prefix("bpm") {
-            let rest = rest.trim();
-            if rest.is_empty() {
-                self.status = "Usage: :bpm <value>  (e.g. :bpm 140)".to_string();
-            } else {
-                match rest.parse::<f32>() {
-                    Ok(bpm) => {
-                        if !bpm.is_finite() {
-                            self.status = format!("Invalid BPM value: '{rest}' — expected a number");
-                        } else {
-                            let clamped = bpm.clamp(20.0, 999.0);
-                            self.record("set BPM");
-                            self.song.bpm = clamped;
-                            self.send_cmd(Command::SetBpm(clamped));
-                            self.set_timed_status(format!("BPM set to {clamped:.1}"));
-                        }
-                    }
-                    Err(_) => {
-                        self.status = format!("Invalid BPM value: '{rest}' — expected a number");
-                    }
-                }
-            }
-        } else if raw.is_empty() {
-            // No-op: Normal mode status bar shows its own fixed hint text.
-        } else {
-            self.status = format!("Unknown command: {raw}");
-        }
-    }
-
-    /// Spawn a background thread to render the full mix and write to `path`.
-    fn start_render_mix(&mut self, path: String) {
-        if self.render_receiver.is_some() {
-            self.status = "Error: render already in progress".to_string();
-            return;
-        }
-        let song = self.song.clone();
-        let progress = Arc::clone(&self.render_progress);
-        progress.store(0, Ordering::Relaxed);
-        let (tx, rx) = std::sync::mpsc::channel();
-        self.render_receiver = Some(rx);
-        self.status = "Rendering mix... 0%".to_string();
-
-        std::thread::spawn(move || {
-            let buffers = load_all_instrument_samples(&song);
-            let result: anyhow::Result<String> = (|| {
-                let audio = vitakt_core::render::render_to_buffer(
-                    &song,
-                    &buffers,
-                    None,
-                    &mut |p| {
-                        progress.store((p * 100.0) as u32, Ordering::Relaxed);
-                    },
-                );
-                write_wav(&path, &audio)?;
-                Ok(format!("Mix exported: {path}"))
-            })();
-            let _ = tx.send(result);
-        });
-    }
-
-    /// Spawn a background thread to render per-track stems into `dir`.
-    fn start_render_stems(&mut self, dir: String) {
-        if self.render_receiver.is_some() {
-            self.status = "Error: render already in progress".to_string();
-            return;
-        }
-        let song = self.song.clone();
-        let progress = Arc::clone(&self.render_progress);
-        progress.store(0, Ordering::Relaxed);
-        let (tx, rx) = std::sync::mpsc::channel();
-        self.render_receiver = Some(rx);
-        self.status = "Rendering stems... 0%".to_string();
-
-        std::thread::spawn(move || {
-            let result: anyhow::Result<String> = (|| {
-                std::fs::create_dir_all(&dir)?;
-                let buffers = load_all_instrument_samples(&song);
-                for track in 0..TRACKS {
-                    let audio = vitakt_core::render::render_to_buffer(
-                        &song,
-                        &buffers,
-                        Some(track),
-                        &mut |p| {
-                            // Scale progress across all TRACKS passes.
-                            let overall = (track as f32 + p) / TRACKS as f32;
-                            progress.store((overall * 100.0) as u32, Ordering::Relaxed);
-                        },
-                    );
-                    let filename = format!("{dir}/track-{:02}.wav", track + 1);
-                    write_wav(&filename, &audio)?;
-                }
-                Ok(format!("Stems exported to: {dir}"))
-            })();
-            let _ = tx.send(result);
-        });
-    }
-
-    /// Poll the render thread receiver; update status bar on completion or progress.
-    fn poll_render(&mut self) {
-        let result = match &self.render_receiver {
-            None => return,
-            Some(rx) => match rx.try_recv() {
-                Ok(r) => r,
-                Err(std::sync::mpsc::TryRecvError::Empty) => {
-                    let pct = self.render_progress.load(Ordering::Relaxed);
-                    self.status = format!("Rendering... {pct}%");
-                    return;
-                }
-                Err(std::sync::mpsc::TryRecvError::Disconnected) => {
-                    Err(anyhow::anyhow!("render thread disconnected unexpectedly"))
-                }
-            },
-        };
-        self.render_receiver = None;
-        match result {
-            Ok(msg) => self.status = msg,
-            Err(e) => self.status = format!("Export error: {e}"),
-        }
-    }
-
-    /// Save a snapshot before a mutation. Call BEFORE any mutation.
-    fn record(&mut self, description: &str) {
-        let snapshot = self.song.clone();
-        self.history.push(description.to_string(), snapshot);
-        self.is_dirty = true;
-    }
-
-    /// Set a status message that clears after 2 seconds.
-    fn set_timed_status(&mut self, msg: String) {
-        self.status = msg;
-        self.status_timer = Some(std::time::Instant::now());
-    }
-
-    /// Undo the most recent mutation.
-    fn do_undo(&mut self) {
-        let current = self.song.clone();
-        if let Some((snapshot, desc)) = self.history.undo(current) {
-            self.song = snapshot;
-            self.is_dirty = true;
-            self.set_timed_status(format!("Undid: {desc}"));
-            self.sync_phrase_to_sequencer();
-            self.sync_song_to_sequencer();
-            self.sync_mixer_to_audio();
-        } else {
-            self.set_timed_status("Nothing to undo".to_string());
-        }
-    }
-
-    /// Redo the most recently undone mutation.
-    fn do_redo(&mut self) {
-        let current = self.song.clone();
-        if let Some((snapshot, desc)) = self.history.redo(current) {
-            self.song = snapshot;
-            self.is_dirty = true;
-            self.set_timed_status(format!("Redid: {desc}"));
-            self.sync_phrase_to_sequencer();
-            self.sync_song_to_sequencer();
-            self.sync_mixer_to_audio();
-        } else {
-            self.set_timed_status("Nothing to redo".to_string());
-        }
-    }
-
-    fn enter_keyboard_mode(&mut self) {
-        self.mode = InputMode::Keyboard;
-    }
-
-    fn exit_keyboard_mode(&mut self) {
-        self.mode = InputMode::Normal;
-    }
-
-    fn keyboard_instrument_prev(&mut self) {
-        if self.keyboard_instrument > 0 {
-            self.keyboard_instrument -= 1;
-        }
-    }
-
-    fn keyboard_instrument_next(&mut self) {
-        if self.keyboard_instrument < 255 {
-            self.keyboard_instrument += 1;
-        }
-    }
-}
-
-/// Slice a decoded sample buffer to the region [sample_start, sample_end] and adjust
-/// loop points to be relative to the new start, clamped to the new buffer length.
-/// Returns the trimmed buffer and adjusted loop points.  `None` values default to
-/// the full buffer / no loop.
-fn apply_sample_bounds(
-    buf: Arc<Vec<f32>>,
-    channels: usize,
-    sample_start: Option<u32>,
-    sample_end: Option<u32>,
-    loop_start: u32,
-    loop_end: u32,
-) -> (Arc<Vec<f32>>, u32, u32) {
-    let total_frames = buf.len() / channels.max(1);
-    let start_frame = sample_start.unwrap_or(0) as usize;
-    let end_frame = sample_end.map(|e| e as usize).unwrap_or(total_frames);
-    let start_frame = start_frame.min(total_frames);
-    let end_frame = end_frame.clamp(start_frame, total_frames);
-    if start_frame == 0 && end_frame == total_frames {
-        return (buf, loop_start, loop_end);
-    }
-    let new_length = end_frame - start_frame;
-    let sliced = Arc::new(buf[start_frame * channels..end_frame * channels].to_vec());
-    let adj_loop_start =
-        ((loop_start as usize).saturating_sub(start_frame)).min(new_length) as u32;
-    let adj_loop_end =
-        ((loop_end as usize).saturating_sub(start_frame)).min(new_length) as u32;
-    (sliced, adj_loop_start, adj_loop_end)
-}
-
-#[cfg(test)]
-mod sample_bounds_tests {
-    use super::*;
-
-    fn make_buf(frames: usize) -> Arc<Vec<f32>> {
-        Arc::new(vec![0.5f32; frames])
-    }
-
-    #[test]
-    fn no_bounds_returns_original_buffer() {
-        let buf = make_buf(100);
-        let (out, ls, le) = apply_sample_bounds(Arc::clone(&buf), 1, None, None, 10, 50);
-        assert_eq!(out.len(), 100);
-        assert_eq!(ls, 10);
-        assert_eq!(le, 50);
-    }
-
-    #[test]
-    fn sample_start_shifts_loop_points() {
-        let buf = make_buf(200);
-        // Slice starts at frame 50; loop was at 60..80 → should become 10..30
-        let (out, ls, le) =
-            apply_sample_bounds(Arc::clone(&buf), 1, Some(50), None, 60, 80);
-        assert_eq!(out.len(), 150);
-        assert_eq!(ls, 10);
-        assert_eq!(le, 30);
-    }
-
-    #[test]
-    fn loop_end_clamped_when_exceeds_sample_end() {
-        let buf = make_buf(500);
-        // sample_end = 200, loop_end = 500 → adj_loop_end must be clamped to 200
-        let (out, ls, le) =
-            apply_sample_bounds(Arc::clone(&buf), 1, Some(0), Some(200), 10, 500);
-        assert_eq!(out.len(), 200);
-        assert_eq!(ls, 10);
-        assert_eq!(le, 200, "loop_end should be clamped to new buffer length");
-    }
-
-    #[test]
-    fn loop_points_before_start_frame_clamped_to_zero() {
-        let buf = make_buf(100);
-        // start_frame = 50, loop_start = 20 (before start) → clamped to 0
-        let (out, ls, le) =
-            apply_sample_bounds(Arc::clone(&buf), 1, Some(50), None, 20, 80);
-        assert_eq!(out.len(), 50);
-        assert_eq!(ls, 0);
-        assert_eq!(le, 30);
-    }
-
-    #[test]
-    fn multichannel_buffer_sliced_correctly() {
-        // 100 stereo frames = 200 samples; slice frames 10..40 = 30 stereo frames = 60 samples
-        let buf = Arc::new(vec![0.5f32; 200]);
-        let (out, _, _) =
-            apply_sample_bounds(Arc::clone(&buf), 2, Some(10), Some(40), 0, 0);
-        assert_eq!(out.len(), 60);
-    }
-}
-
-/// List subdirectories and `.wav` files in `dir`, sorted alphabetically (case-insensitive).
-/// All other file types are excluded.
-fn list_browser_entries(dir: &std::path::Path) -> Vec<BrowserEntry> {
-    list_browser_entries_ext(dir, "wav")
-}
-
-/// List subdirectories and files matching `file_ext` in `dir`, sorted alphabetically
-/// (case-insensitive).  All other file types are excluded.
-fn list_browser_entries_ext(dir: &std::path::Path, file_ext: &str) -> Vec<BrowserEntry> {
-    let mut entries = Vec::new();
-    if let Ok(read_dir) = std::fs::read_dir(dir) {
-        for entry in read_dir.flatten() {
-            let path = entry.path();
-            let name = entry.file_name().to_string_lossy().to_string();
-            if path.is_dir() {
-                entries.push(BrowserEntry::Dir(name));
-            } else if let Some(ext) = path.extension() {
-                if ext.to_ascii_lowercase() == file_ext {
-                    entries.push(BrowserEntry::Wav(name));
-                }
-            }
-        }
-    }
-    entries.sort_by_key(|e| e.sort_key());
-    // Prepend `..` when we're not at the filesystem root.
-    if dir.parent().is_some() {
-        entries.insert(0, BrowserEntry::ParentDir);
-    }
-    entries
-}
-
-// ── Instrument editor field helper ───────────────────────────────────────────
-
-/// Increment or decrement the currently-selected numeric/mode instrument field.
-/// `delta` is +1 or -1.
-fn instr_editor_increment(app: &mut App, delta: i32) {
-    let idx = app.active_instrument;
-    let cursor = app.instr_cursor;
-    if let Some(instr) = app.song.instruments.get_mut(idx) {
-        match cursor {
-            INSTR_FIELD_ROOT => {
-                instr.root_note = (instr.root_note as i32 + delta).clamp(0, 127) as u8;
-                app.sample_root = instr.root_note;
-                let root = instr.root_note;
-                app.send_cmd(Command::SetSampleRoot(root));
-            }
-            INSTR_FIELD_LOOP_START => {
-                let cur = instr.loop_start.unwrap_or(0) as i32;
-                let new_val = (cur + delta).max(0) as u32;
-                instr.loop_start = Some(new_val);
-                let (ls, le) = (new_val, instr.loop_end.unwrap_or(0));
-                app.send_cmd(Command::SetLoopPoints { slot: 0, loop_start: ls, loop_end: le });
-            }
-            INSTR_FIELD_LOOP_END => {
-                let cur = instr.loop_end.unwrap_or(0) as i32;
-                let new_val = (cur + delta).max(0) as u32;
-                instr.loop_end = Some(new_val);
-                let (ls, le) = (instr.loop_start.unwrap_or(0), new_val);
-                app.send_cmd(Command::SetLoopPoints { slot: 0, loop_start: ls, loop_end: le });
-            }
-            INSTR_FIELD_INTERP => {
-                instr.interp_mode = match (&instr.interp_mode, delta > 0) {
-                    (InterpMode::None, true) => InterpMode::Linear,
-                    (InterpMode::Linear, true) => InterpMode::Sinc,
-                    (InterpMode::Sinc, true) => InterpMode::None,
-                    (InterpMode::None, false) => InterpMode::Sinc,
-                    (InterpMode::Linear, false) => InterpMode::None,
-                    (InterpMode::Sinc, false) => InterpMode::Linear,
-                };
-                let mode = instr.interp_mode.clone();
-                app.send_cmd(Command::SetInterpMode { slot: 0, interp_mode: mode });
-            }
-            INSTR_FIELD_VOLUME => {
-                instr.volume = (instr.volume + delta as f32 * 0.05).clamp(0.0, 1.0);
-                // Round to 2 decimal places to avoid float drift
-                instr.volume = (instr.volume * 100.0).round() / 100.0;
-            }
-            INSTR_FIELD_PAN => {
-                instr.pan = (instr.pan + delta as f32 * 0.05).clamp(-1.0, 1.0);
-                instr.pan = (instr.pan * 100.0).round() / 100.0;
-            }
-            _ => {}
-        }
-    }
-}
-
-// ── TUI event loop ────────────────────────────────────────────────────────────
-
-fn run_tui(
-    producer: Option<rtrb::Producer<Command>>,
-    sample_root: u8,
-    seq_playing: Arc<AtomicBool>,
-    current_seq_step: Arc<AtomicU8>,
-    preview_playing: Arc<AtomicBool>,
-    action: CliAction,
-) -> Result<()> {
-    enable_raw_mode()?;
-    let mut stdout = io::stdout();
-    execute!(stdout, EnterAlternateScreen)?;
-    let backend = CrosstermBackend::new(stdout);
-    let mut terminal = Terminal::new(backend)?;
-
-    let mut app = App::new(producer, sample_root, seq_playing, current_seq_step, preview_playing);
-
-    match action {
-        CliAction::ShowStartup => {
-            app.view = View::Startup;
-        }
-        CliAction::OpenFile(path) => {
-            let path_str = path.to_string_lossy().to_string();
-            match storage::load_trk(&path_str) {
-                Ok(song) => {
-                    app.song = vitakt_core::model::migrate(song);
-                    if app.song.phrases.is_empty() {
-                        app.song.phrases.push(vitakt_core::model::Phrase::default());
-                    }
-                    app.sync_phrase_to_sequencer();
-                    app.reload_instruments();
-                    app.sync_song_to_sequencer();
-                    app.sync_mixer_to_audio();
-                    app.set_timed_status(format!("Loaded: {path_str}"));
-                }
-                Err(e) => {
-                    eprintln!("error: could not open '{path_str}': {e}");
-                    disable_raw_mode()?;
-                    return Err(e);
-                }
-            }
-        }
-    }
-
-    loop {
-        // Poll render thread for completion/progress before drawing.
-        app.poll_render();
-
-        // Clear timed status messages after 2 seconds.
-        if let Some(timer) = app.status_timer {
-            if timer.elapsed() >= Duration::from_secs(2) {
-                app.status_timer = None;
-            }
-        }
-
-        // ── Render ────────────────────────────────────────────────────────────
-        terminal.draw(|frame| {
-            let size = frame.area();
-            let outer = Layout::default()
-                .direction(Direction::Vertical)
-                .constraints([Constraint::Min(0), Constraint::Length(1)])
-                .split(size);
-
-            // Main area — depends on active view
-            match app.view {
-                View::Startup => {
-                    let para = render_startup_screen(&app);
-                    frame.render_widget(para, outer[0]);
-                }
-                View::SongView => {
-                    let table = render_song_view(&app);
-                    frame.render_widget(table, outer[0]);
-                }
-                View::ChainView => {
-                    let table = render_chain_view(&app);
-                    frame.render_widget(table, outer[0]);
-                }
-                View::PhraseEditor => {
-                    let phrase = app.phrase();
-                    let seq_playing = app.seq_playing.load(Ordering::Relaxed);
-                    let playback_step = app.current_seq_step.load(Ordering::Relaxed) as usize;
-                    let table = render_phrase_grid(phrase, app.cursor_step, app.cursor_col, app.active_phrase_idx, &app.theme, seq_playing, playback_step);
-                    frame.render_widget(table, outer[0]);
-                }
-                View::InstrumentEditor => {
-                    let para = render_instrument_editor(&app);
-                    frame.render_widget(para, outer[0]);
-                }
-                View::SampleBrowser => {
-                    let para = render_sample_browser(&app, outer[0].height as usize);
-                    frame.render_widget(para, outer[0]);
-                }
-                View::Mixer => {
-                    let table = render_mixer_view(&app);
-                    frame.render_widget(table, outer[0]);
-                }
-            }
-
-            // Quit confirmation modal overlay
-            if matches!(app.mode, InputMode::ConfirmQuit) {
-                let modal_width = 46u16;
-                let modal_height = 4u16;
-                let x = size.width.saturating_sub(modal_width) / 2;
-                let y = size.height.saturating_sub(modal_height) / 2;
-                let modal_area = Rect::new(x, y, modal_width.min(size.width), modal_height.min(size.height));
-                frame.render_widget(Clear, modal_area);
-                let modal = Paragraph::new("Unsaved changes. Quit? (y/n)")
-                    .block(Block::default().borders(Borders::ALL).title(" Confirm Quit "))
-                    .style(Style::default().fg(Color::Yellow).bg(Color::DarkGray));
-                frame.render_widget(modal, modal_area);
-            }
-
-            // Bookmark overlay (shown over the sample browser)
-            if app.browser_show_bookmarks {
-                let valid_bookmarks: Vec<String> = app
-                    .config
-                    .bookmarks
-                    .iter()
-                    .filter(|p| std::path::Path::new(p.as_str()).is_dir())
-                    .cloned()
-                    .collect();
-                let modal_width = (size.width * 2 / 3).max(40).min(size.width);
-                let modal_height = (valid_bookmarks.len() as u16 + 4).min(size.height);
-                let x = size.width.saturating_sub(modal_width) / 2;
-                let y = size.height.saturating_sub(modal_height) / 2;
-                let modal_area = Rect::new(x, y, modal_width, modal_height);
-                frame.render_widget(Clear, modal_area);
-                let mut bm_lines: Vec<ratatui::text::Line<'static>> = Vec::new();
-                for (i, path) in valid_bookmarks.iter().enumerate() {
-                    let (prefix, style) = if i == app.browser_bookmark_cursor {
-                        ("▶ ", Style::default().fg(app.theme.cursor_bg).add_modifier(Modifier::BOLD))
-                    } else {
-                        ("  ", Style::default().fg(Color::White))
-                    };
-                    bm_lines.push(ratatui::text::Line::styled(
-                        format!("{prefix}{path}"),
-                        style,
-                    ));
-                }
-                let bm_para = Paragraph::new(bm_lines)
-                    .block(
-                        Block::default()
-                            .borders(Borders::ALL)
-                            .title(" Bookmarks  [Enter: go  Esc: cancel] ")
-                            .border_style(Style::default().fg(app.theme.screen_title)),
-                    )
-                    .style(Style::default().bg(Color::DarkGray));
-                frame.render_widget(bm_para, modal_area);
-            }
-
-            // Status bar
-            let playing = app.seq_playing.load(Ordering::Relaxed);
-            let seq_step = app.current_seq_step.load(Ordering::Relaxed);
-            let transport = if playing {
-                format!("Step:{:02}  BPM:{:.1}", seq_step, app.song.bpm)
-            } else {
-                format!("Step:{:02}  BPM:{:.1}", seq_step, app.song.bpm)
-            };
-
-            // Mode label: always the leftmost element in the status bar.
-            let mode_label = match app.view {
-                _ if matches!(app.mode, InputMode::Keyboard) => "KEYBOARD",
-                _ if matches!(app.mode, InputMode::ConfirmQuit) => "NORMAL",
-                View::ChainView if app.chain_insert_mode => "INSERT",
-                View::PhraseEditor => match app.mode {
-                    InputMode::Normal => "NORMAL",
-                    InputMode::Insert => "INSERT",
-                    InputMode::Command => "COMMAND",
-                    InputMode::Keyboard => "KEYBOARD",
-                    InputMode::ConfirmQuit => "NORMAL",
-                },
-                _ => "NORMAL",
-            };
-
-            // Insert mode is active when the mode label is INSERT.
-            let insert_active = mode_label == "INSERT";
-            let keyboard_active = mode_label == "KEYBOARD";
-
-            // When a render is in progress, override the status bar with progress.
-            let status_text = if app.render_receiver.is_some() {
-                app.status.clone()
-            } else if matches!(app.mode, InputMode::ConfirmQuit) {
-                "Unsaved changes — y: quit  n/Esc: cancel".to_string()
-            } else if keyboard_active {
-                format!(
-                    "{mode_label}  |  {transport}  |  Ins:{:02}  [/]: change instrument  QWERTY: play note  Esc: normal",
-                    app.keyboard_instrument
-                )
-            } else if app.status_timer.is_some() {
-                format!("{mode_label}  |  {transport}  |  {}", app.status)
-            } else {
-                match app.view {
-                View::Startup => "j/k: navigate  Enter: select  q: quit".to_string(),
-                View::SongView => format!(
-                    "{mode_label}  |  {transport}  |  hjkl: nav  0-9/a-f: chain  Del: clear  Enter: chain view  o: add row below  O: add row above  F3: phrase  ←/→: BPM  q: quit"
-                ),
-                View::ChainView => {
-                    if app.chain_insert_mode {
-                        format!("{mode_label}  |  h/l: phrase ±1  ,/.: transpose ±1  Esc: normal")
-                    } else {
-                        format!("{mode_label}  |  j/k: nav  h/l: phrase  ,/.: transpose  o: add slot below  O: add slot above  d: del slot  Enter: phrase  i: insert  Esc: back")
-                    }
-                }
-                View::PhraseEditor => match app.mode {
-                    InputMode::Normal => format!("{mode_label}  |  {transport}  |  SPC: play  i: insert  Tab: instrument  ←/→: BPM  :: command  q: quit"),
-                    InputMode::Insert => {
-                        let col_hint = match col_to_fx(app.cursor_col) {
-                            Some((s, true)) => {
-                                let buf = &app.fx_edit_buf;
-                                format!("FX{} CMD: [{buf:<3}]  type 3-letter code (VOL/PAN/PIT/RET)", s + 1)
-                            }
-                            Some((s, false)) => {
-                                let buf = &app.fx_edit_buf;
-                                format!("FX{} VAL: [{buf:<3}]  type 0-255, Enter to confirm", s + 1)
-                            }
-                            None if app.cursor_col == COL_NOTE => {
-                                format!("NOTE  Oct:{} Ins:{:02}  QWERTY piano", app.octave, app.active_instrument)
-                            }
-                            None => format!("Col:{} Ins:{:02}", app.cursor_col, app.active_instrument),
-                        };
-                        format!("{mode_label}  |  {transport}  |  {col_hint}  |  Esc: normal")
-                    }
-                    InputMode::Command => format!("{mode_label}  |  {transport}  |  :{}", app.cmd_buf),
-                    // InputMode::Keyboard is handled by the `keyboard_active` branch above
-                    InputMode::Keyboard => unreachable!("Keyboard mode status handled before view match"),
-                    // InputMode::ConfirmQuit is handled by the modal branch above
-                    InputMode::ConfirmQuit => unreachable!("ConfirmQuit status handled before view match"),
-                },
-                View::InstrumentEditor => {
-                    if app.instr_editing {
-                        format!(
-                            "{mode_label}  |  Ins:{:02}  |  Enter: confirm  Esc: cancel",
-                            app.active_instrument
-                        )
-                    } else {
-                        format!(
-                            "{mode_label}  |  Ins:{:02}  |  j/k: nav  h/l: change  i: edit  Enter: browse(sample)  Esc: back",
-                            app.active_instrument
-                        )
-                    }
-                }
-                View::SampleBrowser => {
-                    format!(
-                        "{mode_label}  |  j/k: nav  Enter: select  -/Backspace: up  b: bookmarks  B: bookmark here  Esc: cancel  ({} entries)",
-                        app.browser_entries.len()
-                    )
-                }
-                View::Mixer => {
-                    format!(
-                        "{mode_label}  |  {transport}  |  h/l: track  j/k: field  +/-: adjust  m: mute  s: solo  Esc: back"
-                    )
-                }
-            }};
-
-            let status_bg = if insert_active {
-                app.theme.insert_mode_bg
-            } else if keyboard_active {
-                app.theme.keyboard_mode_bg
-            } else {
-                app.theme.status_bar_bg
-            };
-            let status = Paragraph::new(status_text)
-                .style(Style::default().fg(app.theme.status_bar_fg).bg(status_bg));
-            frame.render_widget(status, outer[1]);
-        })?;
-
-        // ── Input ─────────────────────────────────────────────────────────────
-        if event::poll(Duration::from_millis(16))? {
-            if let Event::Key(key) = event::read()? {
-                // Global undo/redo: works from any view except insert/command mode.
-                let is_insert = matches!((&app.view, &app.mode), (View::PhraseEditor, InputMode::Insert));
-                let is_command = matches!((&app.view, &app.mode), (View::PhraseEditor, InputMode::Command));
-                if !is_insert && !is_command {
-                    if key.code == KeyCode::Char('u') && !key.modifiers.contains(KeyModifiers::CONTROL) {
-                        app.do_undo();
-                        continue;
-                    }
-                    if key.code == KeyCode::Char('r') && key.modifiers.contains(KeyModifiers::CONTROL) {
-                        app.do_redo();
-                        continue;
-                    }
-                }
-                // ── Global: ConfirmQuit modal overrides all per-view key handling ─
-                if matches!(app.mode, InputMode::ConfirmQuit) {
-                    match key.code {
-                        KeyCode::Char('y') | KeyCode::Char('Y') => break,
-                        KeyCode::Char('n') | KeyCode::Char('N') | KeyCode::Esc => {
-                            app.mode = InputMode::Normal;
-                        }
-                        _ => {}
-                    }
-                    continue;
-                }
-
-                // ── Global: Keyboard mode overrides all per-view key handling ─
-                if matches!(app.mode, InputMode::Keyboard) {
-                    match key.code {
-                        KeyCode::Esc => {
-                            app.exit_keyboard_mode();
-                        }
-                        KeyCode::Char('[') => {
-                            app.keyboard_instrument_prev();
-                        }
-                        KeyCode::Char(']') => {
-                            app.keyboard_instrument_next();
-                        }
-                        KeyCode::Char(c) => {
-                            if let Some(semitone) = qwerty_to_semitone(c) {
-                                let base: i32 = 12 * (app.octave as i32 + 1);
-                                let midi = (base + semitone as i32).clamp(0, 127) as u8;
-                                let slot = app.keyboard_instrument as u8;
-                                let root = app.song.instruments
-                                    .get(app.keyboard_instrument)
-                                    .map(|i| i.root_note)
-                                    .unwrap_or(app.sample_root);
-                                let speed = pitch_speed(midi, root);
-                                app.send_cmd(Command::NoteOn { slot, speed });
-                            }
-                        }
-                        _ => {}
-                    }
-                    continue;
-                }
-
-                match app.view {
-                    // ──────────────────────────────────────────────────────────
-                    // Startup screen key handling
-                    // ──────────────────────────────────────────────────────────
-                    View::Startup => match key.code {
-                        KeyCode::Char('q') => {
-                            if app.is_dirty {
-                                app.mode = InputMode::ConfirmQuit;
-                            } else {
-                                break;
-                            }
-                        }
-                        KeyCode::Char('j') | KeyCode::Down => {
-                            app.startup_cursor = (app.startup_cursor + 1) % 2;
-                        }
-                        KeyCode::Char('k') | KeyCode::Up => {
-                            if app.startup_cursor > 0 {
-                                app.startup_cursor -= 1;
-                            }
-                        }
-                        KeyCode::Enter => match app.startup_cursor {
-                            0 => {
-                                // New Project — blank song, go to Song View.
-                                app.view = View::SongView;
-                                app.view_stack.clear();
-                            }
-                            _ => {
-                                // Open File — launch project browser.
-                                app.open_project_browser();
-                            }
-                        },
-                        _ => {}
-                    },
-                    // ──────────────────────────────────────────────────────────
-                    // Song view key handling
-                    // ──────────────────────────────────────────────────────────
-                    View::SongView => match key.code {
-                        KeyCode::Char('q') => {
-                            if app.is_dirty {
-                                app.mode = InputMode::ConfirmQuit;
-                            } else {
-                                break;
-                            }
-                        }
-                        KeyCode::Char(' ') => app.toggle_play(),
-                        KeyCode::F(5) => app.restart_play(),
-                        KeyCode::Left => app.adjust_bpm(-1.0),
-                        KeyCode::Right => app.adjust_bpm(1.0),
-                        KeyCode::Char(':') => {
-                            app.push_view(View::PhraseEditor);
-                            app.mode = InputMode::Command;
-                            app.cmd_buf.clear();
-                        }
-                        // Navigation
-                        KeyCode::Char('j') | KeyCode::Down => {
-                            let rows = app.song.arrangement.len();
-                            if rows == 0 || app.song_cursor_row + 1 >= rows {
-                                // Auto-append a new empty row when j goes past the last
-                                let empty_row = [None; TRACKS];
-                                app.song.arrangement.push(empty_row);
-                            }
-                            app.song_cursor_row = (app.song_cursor_row + 1)
-                                .min(app.song.arrangement.len().saturating_sub(1));
-                        }
-                        KeyCode::Char('k') | KeyCode::Up => {
-                            if app.song_cursor_row > 0 {
-                                app.song_cursor_row -= 1;
-                            }
-                        }
-                        KeyCode::Char('h') => {
-                            if app.song_cursor_track > 0 {
-                                app.song_cursor_track -= 1;
-                            }
-                        }
-                        KeyCode::Char('l') => {
-                            if app.song_cursor_track < TRACKS - 1 {
-                                app.song_cursor_track += 1;
-                            }
-                        }
-                        // Assign chain via hex digit (0-9 → chains 0-9, a-f → chains 10-15)
-                        KeyCode::Char(c) if c.is_ascii_hexdigit() && c != 'j' && c != 'k' && c != 'h' && c != 'l' => {
-                            if let Ok(chain_idx) = u8::from_str_radix(&c.to_string(), 16) {
-                                let row = app.song_cursor_row;
-                                let track = app.song_cursor_track;
-                                app.record(&format!("assign chain to row {} track {}", row, track));
-                                while app.song.arrangement.len() <= row {
-                                    app.song.arrangement.push([None; TRACKS]);
-                                }
-                                app.ensure_chain(chain_idx as usize);
-                                app.song.arrangement[row][track] = Some(chain_idx);
-                                app.sync_song_to_sequencer();
-                            }
-                        }
-                        // Clear cell
-                        KeyCode::Delete | KeyCode::Char('x') => {
-                            let row = app.song_cursor_row;
-                            let track = app.song_cursor_track;
-                            app.record(&format!("clear arrangement row {} track {}", row, track));
-                            if let Some(arr_row) = app.song.arrangement.get_mut(row) {
-                                arr_row[track] = None;
-                                app.sync_song_to_sequencer();
-                            }
-                        }
-                        // Append row below cursor with 'o'
-                        KeyCode::Char('o') => {
-                            app.record("add song row below");
-                            let insert_at = app.song_cursor_row + 1;
-                            app.song.arrangement.insert(insert_at, [None; TRACKS]);
-                        }
-                        // Insert row above cursor with 'O'; cursor follows original row
-                        KeyCode::Char('O') => {
-                            app.record("add song row above");
-                            let insert_at = app.song_cursor_row;
-                            app.song.arrangement.insert(insert_at, [None; TRACKS]);
-                            app.song_cursor_row += 1;
-                        }
-                        // Drill into chain view on Enter
-                        KeyCode::Enter => {
-                            let row = app.song_cursor_row;
-                            let track = app.song_cursor_track;
-                            app.chain_view_track = track;
-                            app.chain_view_row = row;
-                            app.chain_cursor = 0;
-                            app.chain_insert_mode = false;
-                            app.push_view(View::ChainView);
-                        }
-                        // Tab → phrase editor (direct jump)
-                        KeyCode::Tab => {
-                            app.push_view(View::PhraseEditor);
-                            app.mode = InputMode::Normal;
-                        }
-                        // F3 → phrase editor; F4 → instrument editor; F2 → mixer
-                        KeyCode::F(2) => {
-                            app.push_view(View::Mixer);
-                        }
-                        KeyCode::F(3) => {
-                            app.push_view(View::PhraseEditor);
-                        }
-                        KeyCode::F(4) => {
-                            app.push_view(View::InstrumentEditor);
-                        }
-                        KeyCode::Char('/') => {
-                            app.enter_keyboard_mode();
-                        }
-                        _ => {}
-                    },
-
-                    // ──────────────────────────────────────────────────────────
-                    // Chain view key handling
-                    // ──────────────────────────────────────────────────────────
-                    View::ChainView => {
-                        let track = app.chain_view_track;
-                        let row = app.chain_view_row;
-                        let chain_idx_opt = app.song.arrangement
-                            .get(row)
-                            .and_then(|r| r[track])
-                            .map(|c| c as usize);
-
-                        if app.chain_insert_mode {
-                            match key.code {
-                                KeyCode::Esc => {
-                                    app.chain_insert_mode = false;
-                                }
-                                // h/l: adjust phrase index
-                                KeyCode::Char('h') | KeyCode::Left => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        let cursor = app.chain_cursor;
-                                        if cursor < app.song.chains[ci].slots.len() {
-                                            app.record("adjust chain phrase");
-                                            let max = app.song.phrases.len().saturating_sub(1) as u8;
-                                            app.song.chains[ci].slots[cursor].phrase =
-                                                app.song.chains[ci].slots[cursor].phrase.saturating_sub(1).min(max);
-                                            app.sync_song_to_sequencer();
-                                        }
-                                    }
-                                }
-                                KeyCode::Char('l') | KeyCode::Right => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        let cursor = app.chain_cursor;
-                                        if cursor < app.song.chains[ci].slots.len() {
-                                            app.record("adjust chain phrase");
-                                            let max = app.song.phrases.len().saturating_sub(1) as u8;
-                                            app.song.chains[ci].slots[cursor].phrase =
-                                                (app.song.chains[ci].slots[cursor].phrase + 1).min(max);
-                                            app.sync_song_to_sequencer();
-                                        }
-                                    }
-                                }
-                                // ,/. : adjust transpose (semitones down/up)
-                                KeyCode::Char(',') => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        let cursor = app.chain_cursor;
-                                        if cursor < app.song.chains[ci].slots.len() {
-                                            app.record("adjust chain transpose");
-                                            app.song.chains[ci].slots[cursor].transpose =
-                                                app.song.chains[ci].slots[cursor].transpose.saturating_sub(1);
-                                            app.sync_song_to_sequencer();
-                                        }
-                                    }
-                                }
-                                KeyCode::Char('.') => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        let cursor = app.chain_cursor;
-                                        if cursor < app.song.chains[ci].slots.len() {
-                                            app.record("adjust chain transpose");
-                                            app.song.chains[ci].slots[cursor].transpose =
-                                                app.song.chains[ci].slots[cursor].transpose.saturating_add(1);
-                                            app.sync_song_to_sequencer();
-                                        }
-                                    }
-                                }
-                                _ => {}
-                            }
-                        } else {
-                            // Normal mode
-                            match key.code {
-                                KeyCode::Esc | KeyCode::Backspace => app.pop_view(),
-                                KeyCode::Char('j') | KeyCode::Down => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        let len = app.song.chains[ci].slots.len();
-                                        if len > 0 {
-                                            app.chain_cursor = (app.chain_cursor + 1) % len;
-                                        }
-                                    }
-                                }
-                                KeyCode::Char('k') | KeyCode::Up => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        let len = app.song.chains[ci].slots.len();
-                                        if len > 0 {
-                                            app.chain_cursor = (app.chain_cursor + len - 1) % len;
-                                        }
-                                    }
-                                }
-                                // h/l: adjust phrase index in normal mode too
-                                KeyCode::Char('h') | KeyCode::Left => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        let cursor = app.chain_cursor;
-                                        if cursor < app.song.chains[ci].slots.len() {
-                                            app.record("adjust chain phrase");
-                                            app.song.chains[ci].slots[cursor].phrase =
-                                                app.song.chains[ci].slots[cursor].phrase.saturating_sub(1);
-                                            app.sync_song_to_sequencer();
-                                        }
-                                    }
-                                }
-                                KeyCode::Char('l') | KeyCode::Right => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        let cursor = app.chain_cursor;
-                                        if cursor < app.song.chains[ci].slots.len() {
-                                            app.record("adjust chain phrase");
-                                            let max = app.song.phrases.len().saturating_sub(1) as u8;
-                                            app.song.chains[ci].slots[cursor].phrase =
-                                                (app.song.chains[ci].slots[cursor].phrase + 1).min(max);
-                                            app.sync_song_to_sequencer();
-                                        }
-                                    }
-                                }
-                                // ,/. : adjust transpose
-                                KeyCode::Char(',') => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        let cursor = app.chain_cursor;
-                                        if cursor < app.song.chains[ci].slots.len() {
-                                            app.record("adjust chain transpose");
-                                            app.song.chains[ci].slots[cursor].transpose =
-                                                app.song.chains[ci].slots[cursor].transpose.saturating_sub(1);
-                                            app.sync_song_to_sequencer();
-                                        }
-                                    }
-                                }
-                                KeyCode::Char('.') => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        let cursor = app.chain_cursor;
-                                        if cursor < app.song.chains[ci].slots.len() {
-                                            app.record("adjust chain transpose");
-                                            app.song.chains[ci].slots[cursor].transpose =
-                                                app.song.chains[ci].slots[cursor].transpose.saturating_add(1);
-                                            app.sync_song_to_sequencer();
-                                        }
-                                    }
-                                }
-                                // i: enter insert mode
-                                KeyCode::Char('i') => {
-                                    if chain_idx_opt.is_some() {
-                                        app.chain_insert_mode = true;
-                                    }
-                                }
-                                // o: insert slot below cursor
-                                KeyCode::Char('o') => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        app.record("add chain slot below");
-                                        let insert_at = (app.chain_cursor + 1).min(app.song.chains[ci].slots.len());
-                                        app.song.chains[ci].slots.insert(insert_at, ChainSlot {
-                                            phrase: 0,
-                                            transpose: 0,
-                                        });
-                                        app.chain_cursor = insert_at;
-                                        app.sync_song_to_sequencer();
-                                    }
-                                }
-                                // O: insert slot above cursor; cursor follows original slot
-                                KeyCode::Char('O') => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        app.record("add chain slot above");
-                                        let insert_at = app.chain_cursor;
-                                        app.song.chains[ci].slots.insert(insert_at, ChainSlot {
-                                            phrase: 0,
-                                            transpose: 0,
-                                        });
-                                        app.chain_cursor += 1;
-                                        app.sync_song_to_sequencer();
-                                    }
-                                }
-                                // d: delete current slot
-                                KeyCode::Char('d') | KeyCode::Delete => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        let len = app.song.chains[ci].slots.len();
-                                        if len > 0 {
-                                            app.record("delete chain slot");
-                                            app.song.chains[ci].slots.remove(app.chain_cursor);
-                                            if app.chain_cursor >= app.song.chains[ci].slots.len() && app.chain_cursor > 0 {
-                                                app.chain_cursor -= 1;
-                                            }
-                                            app.sync_song_to_sequencer();
-                                        }
-                                    }
-                                }
-                                // Enter: drill into PhraseEditor for selected slot
-                                KeyCode::Enter => {
-                                    if let Some(ci) = chain_idx_opt {
-                                        if let Some(slot) = app.song.chains[ci].slots.get(app.chain_cursor) {
-                                            let phrase_idx = slot.phrase as usize;
-                                            while app.song.phrases.len() <= phrase_idx {
-                                                app.song.phrases.push(vitakt_core::model::Phrase::default());
-                                            }
-                                            app.active_phrase_idx = phrase_idx;
-                                            app.push_view(View::PhraseEditor);
-                                            app.mode = InputMode::Normal;
-                                        }
-                                    }
-                                }
-                                // F1: song view (back), F3: phrase editor
-                                KeyCode::F(1) => app.pop_view(),
-                                KeyCode::F(3) => {
-                                    app.push_view(View::PhraseEditor);
-                                    app.mode = InputMode::Normal;
-                                }
-                                KeyCode::Char('/') => {
-                                    app.enter_keyboard_mode();
-                                }
-                                _ => {}
-                            }
-                        }
-                    },
-
-                    // ──────────────────────────────────────────────────────────
-                    // Phrase editor key handling
-                    // ──────────────────────────────────────────────────────────
-                    View::PhraseEditor => match app.mode {
-                        // ── Normal mode ───────────────────────────────────────
-                        InputMode::Normal => {
-                            app.yy_pending = false;
-                            match key.code {
-                                KeyCode::Esc => app.pop_view(),
-                                KeyCode::Char('q') => {
-                                    if app.is_dirty {
-                                        app.mode = InputMode::ConfirmQuit;
-                                    } else {
-                                        break;
-                                    }
-                                }
-                                KeyCode::F(1) => {
-                                    // Go to Song View (top)
-                                    app.view_stack.clear();
-                                    app.view = View::SongView;
-                                }
-                                KeyCode::Char(':') => {
-                                    app.mode = InputMode::Command;
-                                    app.cmd_buf.clear();
-                                }
-                                KeyCode::Char('i') => {
-                                    app.mode = InputMode::Insert;
-                                }
-                                // Tab → instrument editor
-                                KeyCode::Tab => app.open_instrument_editor(),
-                                // Transport: Space = play/stop, F5 = restart from 0
-                                KeyCode::Char(' ') => app.toggle_play(),
-                                KeyCode::F(5) => app.restart_play(),
-                                // BPM adjustment: left/right arrows
-                                KeyCode::Left => app.adjust_bpm(-1.0),
-                                KeyCode::Right => app.adjust_bpm(1.0),
-                                // Vim navigation (up/down)
-                                KeyCode::Char('j') | KeyCode::Down => {
-                                    app.cursor_step =
-                                        (app.cursor_step + 1) % STEPS_PER_PHRASE;
-                                }
-                                KeyCode::Char('k') | KeyCode::Up => {
-                                    app.cursor_step =
-                                        (app.cursor_step + STEPS_PER_PHRASE - 1)
-                                            % STEPS_PER_PHRASE;
-                                }
-                                KeyCode::Char('h') => {
-                                    if app.cursor_col > 0 {
-                                        app.cursor_col -= 1;
-                                    }
-                                    app.fx_edit_buf.clear();
-                                }
-                                KeyCode::Char('l') => {
-                                    if app.cursor_col + 1 < COL_COUNT {
-                                        app.cursor_col += 1;
-                                    }
-                                    app.fx_edit_buf.clear();
-                                }
-                                // Delete / clear step or active FX slot
-                                KeyCode::Char('d') | KeyCode::Delete => {
-                                    let idx = app.cursor_step;
-                                    if let Some((slot_idx, _)) = col_to_fx(app.cursor_col) {
-                                        app.record(&format!("clear FX slot {} at step {}", slot_idx + 1, app.cursor_step));
-                                        // Clear just the active FX slot.
-                                        app.phrase_mut().steps[idx].fx[slot_idx] =
-                                            vitakt_core::model::FxSlot::default();
-                                    } else {
-                                        app.record(&format!("clear step {}", app.cursor_step));
-                                        // Clear the entire step.
-                                        let step = &mut app.phrase_mut().steps[idx];
-                                        step.note = None;
-                                        step.instrument = None;
-                                        step.velocity = 0;
-                                        step.fx = Default::default();
-                                    }
-                                    app.fx_edit_buf.clear();
-                                    app.sync_phrase_to_sequencer();
-                                }
-                                // yy — copy step
-                                KeyCode::Char('y') => {
-                                    if app.yy_pending {
-                                        app.yanked_step =
-                                            Some(app.phrase().steps[app.cursor_step].clone());
-                                        app.set_timed_status(format!(
-                                            "Yanked step {}",
-                                            app.cursor_step
-                                        ));
-                                        app.yy_pending = false;
-                                    } else {
-                                        app.yy_pending = true;
-                                        continue; // don't reset yy_pending below
-                                    }
-                                }
-                                // p — paste step
-                                KeyCode::Char('p') => {
-                                    if let Some(s) = app.yanked_step.clone() {
-                                        let idx = app.cursor_step;
-                                        app.record(&format!("paste at step {}", app.cursor_step));
-                                        app.phrase_mut().steps[idx] = s;
-                                        app.sync_phrase_to_sequencer();
-                                    }
-                                }
-                                KeyCode::Char('/') => {
-                                    app.enter_keyboard_mode();
-                                }
-                                _ => {}
-                            }
-                            if !matches!(key.code, KeyCode::Char('y')) {
-                                app.yy_pending = false;
-                            }
-                        }
-
-                        // ── Insert mode ───────────────────────────────────────
-                        InputMode::Insert => match key.code {
-                            KeyCode::Esc => {
-                                app.mode = InputMode::Normal;
-                                app.fx_edit_buf.clear();
-                            }
-                            KeyCode::Up => {
-                                app.cursor_step =
-                                    (app.cursor_step + STEPS_PER_PHRASE - 1)
-                                        % STEPS_PER_PHRASE;
-                                app.fx_edit_buf.clear();
-                            }
-                            KeyCode::Down => {
-                                app.cursor_step =
-                                    (app.cursor_step + 1) % STEPS_PER_PHRASE;
-                                app.fx_edit_buf.clear();
-                            }
-                            KeyCode::Left => {
-                                if app.cursor_col > 0 {
-                                    app.cursor_col -= 1;
-                                }
-                                app.fx_edit_buf.clear();
-                            }
-                            KeyCode::Right => {
-                                if app.cursor_col + 1 < COL_COUNT {
-                                    app.cursor_col += 1;
-                                }
-                                app.fx_edit_buf.clear();
-                            }
-                            // Octave up/down only active on the NOTE column.
-                            KeyCode::Char('-') if app.cursor_col == COL_NOTE => {
-                                if app.octave > 1 {
-                                    app.octave -= 1;
-                                }
-                            }
-                            KeyCode::Char('=') if app.cursor_col == COL_NOTE => {
-                                if app.octave < 8 {
-                                    app.octave += 1;
-                                }
-                            }
-                            // Delete / Backspace clears active FX field.
-                            KeyCode::Delete | KeyCode::Backspace
-                                if col_to_fx(app.cursor_col).is_some() =>
-                            {
-                                if !app.fx_edit_buf.is_empty() {
-                                    app.fx_edit_buf.pop();
-                                } else if let Some((slot_idx, _)) = col_to_fx(app.cursor_col) {
-                                    let step_idx = app.cursor_step;
-                                    app.record(&format!("clear FX slot {} at step {}", slot_idx + 1, app.cursor_step));
-                                    app.phrase_mut().steps[step_idx].fx[slot_idx] =
-                                        vitakt_core::model::FxSlot::default();
-                                    app.sync_phrase_to_sequencer();
-                                }
-                            }
-                            KeyCode::Char(c) => {
-                                match col_to_fx(app.cursor_col) {
-                                    Some((slot_idx, true)) => {
-                                        // FX command field: accumulate up to 3 alpha chars.
-                                        if c.is_alphabetic() {
-                                            app.fx_edit_buf.push(c.to_ascii_uppercase());
-                                            if app.fx_edit_buf.len() == 3 {
-                                                let buf = app.fx_edit_buf.clone();
-                                                app.fx_edit_buf.clear();
-                                                let step_idx = app.cursor_step;
-                                                if let Some(cmd) = FxCommand::from_code(&buf) {
-                                                    app.record(&format!("set FX{} command at step {}", slot_idx + 1, app.cursor_step));
-                                                    app.phrase_mut().steps[step_idx].fx[slot_idx]
-                                                        .command = cmd.id();
-                                                    app.sync_phrase_to_sequencer();
-                                                    // Advance to value column.
-                                                    if app.cursor_col + 1 < COL_COUNT {
-                                                        app.cursor_col += 1;
-                                                    }
-                                                } else {
-                                                    app.status =
-                                                        format!("Unknown FX command: {buf}");
-                                                }
-                                            }
-                                        }
-                                    }
-                                    Some((slot_idx, false)) => {
-                                        // FX value field: accumulate up to 3 decimal digits.
-                                        if c.is_ascii_digit() {
-                                            app.fx_edit_buf.push(c);
-                                            if app.fx_edit_buf.len() == 3 {
-                                                let buf = app.fx_edit_buf.clone();
-                                                app.fx_edit_buf.clear();
-                                                if let Ok(v) = buf.parse::<u16>() {
-                                                    let step_idx = app.cursor_step;
-                                                    app.record(&format!("set FX{} value at step {}", slot_idx + 1, app.cursor_step));
-                                                    app.phrase_mut().steps[step_idx].fx[slot_idx]
-                                                        .value = v.min(255) as u8;
-                                                    app.sync_phrase_to_sequencer();
-                                                    // Advance cursor to next step.
-                                                    app.cursor_step =
-                                                        (app.cursor_step + 1) % STEPS_PER_PHRASE;
-                                                }
-                                            }
-                                        }
-                                    }
-                                    None => {
-                                        // NOTE column: QWERTY piano.
-                                        if app.cursor_col == COL_NOTE {
-                                            if let Some(semitone) = qwerty_to_semitone(c) {
-                                                let base: i32 = 12 * (app.octave as i32 + 1);
-                                                let midi =
-                                                    (base + semitone as i32).clamp(0, 127) as u8;
-                                                app.enter_note(midi);
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                            // Enter confirms FX value entry early (before 3 digits).
-                            KeyCode::Enter if col_to_fx(app.cursor_col).is_some() => {
-                                if let Some((slot_idx, false)) = col_to_fx(app.cursor_col) {
-                                    let buf = app.fx_edit_buf.clone();
-                                    app.fx_edit_buf.clear();
-                                    if !buf.is_empty() {
-                                        if let Ok(v) = buf.parse::<u16>() {
-                                            let step_idx = app.cursor_step;
-                                            app.record(&format!("set FX value at step {}", app.cursor_step));
-                                            app.phrase_mut().steps[step_idx].fx[slot_idx].value =
-                                                v.min(255) as u8;
-                                            app.sync_phrase_to_sequencer();
-                                            app.cursor_step =
-                                                (app.cursor_step + 1) % STEPS_PER_PHRASE;
-                                        }
-                                    }
-                                }
-                            }
-                            _ => {}
-                        },
-
-                        // ── Command mode ──────────────────────────────────────
-                        InputMode::Command => match key.code {
-                            KeyCode::Enter => app.execute_command(),
-                            KeyCode::Esc => {
-                                app.mode = InputMode::Normal;
-                                app.cmd_buf.clear();
-                            }
-                            KeyCode::Backspace => {
-                                app.cmd_buf.pop();
-                            }
-                            KeyCode::Char(c) => {
-                                app.cmd_buf.push(c);
-                            }
-                            _ => {}
-                        },
-                        // Keyboard mode is handled globally before this match; this arm
-                        // is unreachable but required for exhaustiveness.
-                        InputMode::Keyboard => {}
-                        // ConfirmQuit is handled globally before this match; this arm
-                        // is unreachable but required for exhaustiveness.
-                        InputMode::ConfirmQuit => {}
-                    },
-
-                    // ──────────────────────────────────────────────────────────
-                    // Instrument editor key handling
-                    // ──────────────────────────────────────────────────────────
-                    View::InstrumentEditor => {
-                        if app.instr_editing {
-                            // Text-edit sub-mode for Name / Sample fields
-                            match key.code {
-                                KeyCode::Enter => {
-                                    let buf = app.instr_edit_buf.clone();
-                                    let cursor = app.instr_cursor;
-                                    app.record(&format!("edit instrument {} name/sample", app.active_instrument));
-                                    if let Some(instr) =
-                                        app.song.instruments.get_mut(app.active_instrument)
-                                    {
-                                        match cursor {
-                                            INSTR_FIELD_NAME => instr.name = buf,
-                                            INSTR_FIELD_SAMPLE => {
-                                                instr.sample = Some(
-                                                    vitakt_core::model::Sample::from_path(&buf),
-                                                );
-                                            }
-                                            _ => {}
-                                        }
-                                    }
-                                    app.instr_editing = false;
-                                    app.instr_edit_buf.clear();
-                                    if cursor == INSTR_FIELD_SAMPLE {
-                                        app.reload_instrument_sample();
-                                    }
-                                }
-                                KeyCode::Esc => {
-                                    app.instr_editing = false;
-                                    app.instr_edit_buf.clear();
-                                }
-                                KeyCode::Backspace => {
-                                    app.instr_edit_buf.pop();
-                                }
-                                KeyCode::Char(c) => {
-                                    app.instr_edit_buf.push(c);
-                                }
-                                _ => {}
-                            }
-                        } else {
-                            // Normal instrument editor navigation
-                            match key.code {
-                                KeyCode::Esc => {
-                                    app.pop_view();
-                                }
-                                KeyCode::Char('j') | KeyCode::Down => {
-                                    app.instr_cursor =
-                                        (app.instr_cursor + 1) % INSTR_FIELD_COUNT;
-                                }
-                                KeyCode::Char('k') | KeyCode::Up => {
-                                    app.instr_cursor =
-                                        (app.instr_cursor + INSTR_FIELD_COUNT - 1)
-                                            % INSTR_FIELD_COUNT;
-                                }
-                                // Enter text edit mode (for name/sample) or open browser
-                                KeyCode::Char('i') => {
-                                    let cursor = app.instr_cursor;
-                                    if cursor == INSTR_FIELD_SAMPLE {
-                                        // Open browser directly on sample field
-                                        app.open_sample_browser();
-                                    } else if cursor == INSTR_FIELD_NAME {
-                                        // Pre-fill with current name
-                                        let cur_name = app
-                                            .song
-                                            .instruments
-                                            .get(app.active_instrument)
-                                            .map(|i| i.name.clone())
-                                            .unwrap_or_default();
-                                        app.instr_edit_buf = cur_name;
-                                        app.instr_editing = true;
-                                    } else {
-                                        // For non-text fields, treat 'i' like 'l' (increment)
-                                        app.record(&format!("edit instrument {} field {}", app.active_instrument, app.instr_cursor));
-                                        instr_editor_increment(&mut app, 1);
-                                    }
-                                }
-                                // Enter opens browser on sample field
-                                KeyCode::Enter => {
-                                    if app.instr_cursor == INSTR_FIELD_SAMPLE {
-                                        app.open_sample_browser();
-                                    }
-                                }
-                                // h/l or Left/Right adjust numeric/mode fields
-                                KeyCode::Char('h') | KeyCode::Left => {
-                                    app.record(&format!("edit instrument {} field {}", app.active_instrument, app.instr_cursor));
-                                    instr_editor_increment(&mut app, -1);
-                                }
-                                KeyCode::Char('l') | KeyCode::Right => {
-                                    app.record(&format!("edit instrument {} field {}", app.active_instrument, app.instr_cursor));
-                                    instr_editor_increment(&mut app, 1);
-                                }
-                                _ => {}
-                            }
-                        }
-                    }
-
-                    // ──────────────────────────────────────────────────────────
-                    // Sample browser key handling
-                    // ──────────────────────────────────────────────────────────
-                    View::SampleBrowser => {
-                        // When bookmark overlay is open, intercept all keys for it.
-                        if app.browser_show_bookmarks {
-                            let valid_len = app
-                                .config
-                                .bookmarks
-                                .iter()
-                                .filter(|p| std::path::Path::new(p.as_str()).is_dir())
-                                .count();
-                            match key.code {
-                                KeyCode::Esc => {
-                                    app.browser_overlay_esc();
-                                }
-                                KeyCode::Char('j') | KeyCode::Down => {
-                                    if valid_len > 0 {
-                                        app.browser_bookmark_cursor =
-                                            (app.browser_bookmark_cursor + 1) % valid_len;
-                                    }
-                                }
-                                KeyCode::Char('k') | KeyCode::Up => {
-                                    if valid_len > 0 {
-                                        app.browser_bookmark_cursor =
-                                            (app.browser_bookmark_cursor + valid_len - 1)
-                                                % valid_len;
-                                    }
-                                }
-                                KeyCode::Enter => {
-                                    app.browser_overlay_enter();
-                                }
-                                _ => {}
-                            }
-                        } else {
-                            match key.code {
-                                KeyCode::Esc => {
-                                    app.pop_view();
-                                }
-                                KeyCode::Char(' ') => app.browser_preview_toggle(),
-                                KeyCode::Char('j') | KeyCode::Down => {
-                                    if !app.browser_entries.is_empty() {
-                                        app.browser_cursor =
-                                            (app.browser_cursor + 1) % app.browser_entries.len();
-                                        let available = terminal
-                                            .size()
-                                            .map(|r| (r.height as usize).saturating_sub(5))
-                                            .unwrap_or(0);
-                                        app.browser_clamp_scroll(available);
-                                    }
-                                }
-                                KeyCode::Char('k') | KeyCode::Up => {
-                                    if !app.browser_entries.is_empty() {
-                                        app.browser_cursor =
-                                            (app.browser_cursor + app.browser_entries.len() - 1)
-                                                % app.browser_entries.len();
-                                        let available = terminal
-                                            .size()
-                                            .map(|r| (r.height as usize).saturating_sub(5))
-                                            .unwrap_or(0);
-                                        app.browser_clamp_scroll(available);
-                                    }
-                                }
-                                KeyCode::Enter => app.browser_enter(),
-                                KeyCode::Backspace | KeyCode::Char('-') => app.browser_go_up(),
-                                KeyCode::Char('b') => {
-                                    app.browser_try_open_bookmarks();
-                                }
-                                KeyCode::Char('B') => {
-                                    app.browser_add_bookmark();
-                                }
-                                _ => {}
-                            }
-                        }
-                    }
-
-                    // ──────────────────────────────────────────────────────────
-                    // Mixer view key handling
-                    // ──────────────────────────────────────────────────────────
-                    View::Mixer => match key.code {
-                        KeyCode::Esc | KeyCode::F(2) => app.pop_view(),
-
-                        // Track navigation (left/right)
-                        KeyCode::Char('h') | KeyCode::Left => {
-                            if app.mixer_cursor_track > 0 {
-                                app.mixer_cursor_track -= 1;
-                            }
-                        }
-                        KeyCode::Char('l') | KeyCode::Right => {
-                            if app.mixer_cursor_track < TRACKS - 1 {
-                                app.mixer_cursor_track += 1;
-                            }
-                        }
-
-                        // Field navigation (up/down)
-                        KeyCode::Char('j') | KeyCode::Down => {
-                            app.mixer_cursor_field =
-                                (app.mixer_cursor_field + 1) % MIXER_FIELD_COUNT;
-                        }
-                        KeyCode::Char('k') | KeyCode::Up => {
-                            app.mixer_cursor_field =
-                                (app.mixer_cursor_field + MIXER_FIELD_COUNT - 1)
-                                    % MIXER_FIELD_COUNT;
-                        }
-
-                        // Increment/decrement numeric fields (+/=  and  -)
-                        KeyCode::Char('+') | KeyCode::Char('=') => {
-                            app.record(&format!("set mixer track {} field", app.mixer_cursor_track));
-                            let t = app.mixer_cursor_track;
-                            let m = &mut app.song.mixer[t];
-                            match app.mixer_cursor_field {
-                                MIXER_FIELD_VOL => {
-                                    m.volume = ((m.volume + 0.05) * 100.0).round() / 100.0;
-                                    m.volume = m.volume.clamp(0.0, 2.0);
-                                    let v = m.volume;
-                                    app.send_cmd(Command::SetTrackVolume {
-                                        track: t as u8,
-                                        volume: v,
-                                    });
-                                }
-                                MIXER_FIELD_PAN => {
-                                    m.pan = ((m.pan + 0.05) * 100.0).round() / 100.0;
-                                    m.pan = m.pan.clamp(-1.0, 1.0);
-                                    let p = m.pan;
-                                    app.send_cmd(Command::SetTrackPan {
-                                        track: t as u8,
-                                        pan: p,
-                                    });
-                                }
-                                MIXER_FIELD_SEND => {
-                                    m.fx_send = ((m.fx_send + 0.05) * 100.0).round() / 100.0;
-                                    m.fx_send = m.fx_send.clamp(0.0, 1.0);
-                                }
-                                _ => {}
-                            }
-                        }
-                        KeyCode::Char('-') => {
-                            app.record(&format!("set mixer track {} field", app.mixer_cursor_track));
-                            let t = app.mixer_cursor_track;
-                            let m = &mut app.song.mixer[t];
-                            match app.mixer_cursor_field {
-                                MIXER_FIELD_VOL => {
-                                    m.volume = ((m.volume - 0.05) * 100.0).round() / 100.0;
-                                    m.volume = m.volume.clamp(0.0, 2.0);
-                                    let v = m.volume;
-                                    app.send_cmd(Command::SetTrackVolume {
-                                        track: t as u8,
-                                        volume: v,
-                                    });
-                                }
-                                MIXER_FIELD_PAN => {
-                                    m.pan = ((m.pan - 0.05) * 100.0).round() / 100.0;
-                                    m.pan = m.pan.clamp(-1.0, 1.0);
-                                    let p = m.pan;
-                                    app.send_cmd(Command::SetTrackPan {
-                                        track: t as u8,
-                                        pan: p,
-                                    });
-                                }
-                                MIXER_FIELD_SEND => {
-                                    m.fx_send = ((m.fx_send - 0.05) * 100.0).round() / 100.0;
-                                    m.fx_send = m.fx_send.clamp(0.0, 1.0);
-                                }
-                                _ => {}
-                            }
-                        }
-
-                        // Toggle mute (m key or Enter on MUTE row)
-                        KeyCode::Char('m') => {
-                            let t = app.mixer_cursor_track;
-                            app.record(&format!("toggle mute track {}", t));
-                            let new_mute = !app.song.mixer[t].mute;
-                            app.song.mixer[t].mute = new_mute;
-                            app.send_cmd(Command::SetTrackMute {
-                                track: t as u8,
-                                mute: new_mute,
-                            });
-                        }
-
-                        // Toggle solo (s key or Enter on SOLO row)
-                        KeyCode::Char('s') => {
-                            let t = app.mixer_cursor_track;
-                            app.record(&format!("toggle solo track {}", t));
-                            let new_solo = !app.song.mixer[t].solo;
-                            app.song.mixer[t].solo = new_solo;
-                            app.send_cmd(Command::SetTrackSolo {
-                                track: t as u8,
-                                active: new_solo,
-                            });
-                        }
-
-                        // Enter toggles the current field (mute/solo rows) or steps on others
-                        KeyCode::Enter => {
-                            let t = app.mixer_cursor_track;
-                            match app.mixer_cursor_field {
-                                MIXER_FIELD_MUTE => {
-                                    app.record(&format!("toggle mute track {}", t));
-                                    let new_mute = !app.song.mixer[t].mute;
-                                    app.song.mixer[t].mute = new_mute;
-                                    app.send_cmd(Command::SetTrackMute {
-                                        track: t as u8,
-                                        mute: new_mute,
-                                    });
-                                }
-                                MIXER_FIELD_SOLO => {
-                                    app.record(&format!("toggle solo track {}", t));
-                                    let new_solo = !app.song.mixer[t].solo;
-                                    app.song.mixer[t].solo = new_solo;
-                                    app.send_cmd(Command::SetTrackSolo {
-                                        track: t as u8,
-                                        active: new_solo,
-                                    });
-                                }
-                                _ => {}
-                            }
-                        }
-
-                        // Transport controls pass through even from Mixer
-                        KeyCode::Char(' ') => app.toggle_play(),
-                        KeyCode::F(5) => app.restart_play(),
-
-                        _ => {}
-                    },
-                }
-            }
-        }
-    }
-
-    disable_raw_mode()?;
-    execute!(terminal.backend_mut(), LeaveAlternateScreen)?;
-    terminal.show_cursor()?;
-    Ok(())
-}
 
 // ── Entry point ───────────────────────────────────────────────────────────────
 
@@ -2299,7 +69,7 @@ fn main() -> Result<()> {
         Arc::clone(&seq_playing),
         Arc::clone(&current_seq_step),
         Arc::clone(&preview_playing),
-        None, // samples are loaded from project instruments, not the CLI
+        None,
         initial_bpm,
     )
     .unwrap_or_else(|e| {
@@ -2307,11 +77,10 @@ fn main() -> Result<()> {
         panic!("audio unavailable: {e}")
     });
 
-    run_tui(Some(producer), 60, seq_playing, current_seq_step, preview_playing, action)?;
+    tui::run_tui(Some(producer), 60, seq_playing, current_seq_step, preview_playing, action)?;
     Ok(())
 }
 
-// ── Helpers ───────────────────────────────────────────────────────────────────
 
 #[cfg(test)]
 mod tests {
